Fixes:    QIODevice getChar() breakage when reading 0xff characters. AutoTest: 204563 Details:  This exposed a bug in thin - done with Sam.
[git-p4: depot-paths = "//depot/qt/main/": change = 204566]

+++ src/corelib/tools/qringbuffer_p.h
@@ -143,7 +143,7 @@ public:
             return -1;
         char c = *readPointer();
         free(1);
-        return c;
+        return int(uchar(c));
     }
 
     inline void putChar(char c) {
